Add support for OpenSSL
We can now build two implementations: one based on GnuTLS and one based
on OpenSSL. The latter doesn't work yet, but is pretty close to working.

The goal here is that we want to support external APIs which can work
with both the major (and FIPS certified) crypto libraries. Behavior with
both backends should be the same.